+++ client/src/components/Domain/Files/FilesEventsDetailContainer/FilesEventsDetailContainer.module.scss
@@ -1,3 +1,6 @@
+@import 'styles/variables/colors';
+@import 'styles/variables/spacing';
+
 .container {
     width: 403px;
     height: 700px;
